test(downloader): fold branch/http extra tests into siblings
Continue the test-file consolidation. The two remaining standalone
files were the same white-box package and same source-file topic as
existing files, so fold them in (no test logic changed):

  cache_branch_test -> cache_extra_test   (cache.go white-box tests)
  http_extra_test   -> http_test          (http.go helper tests)

Downloader package goes from 7 test files to 5. All 138 test/helper
functions across both packages are preserved.
